add router for changing pages
