+++ src/pages/goalpage/goalpage.js
@@ -1,0 +1,3 @@
+export default ()=>{
+  return console.log('dfdfdffdfdf');
+}
