+++ src/pages/goalpage/index.hbs
@@ -1,0 +1,1 @@
+  <p class='pt-64'>sdfsd</p>
