+++ src/pages/homepage/index.hbs
@@ -1,6 +1,4 @@
-<main class="bg-gray-50 min-h-screen pt-14 flex flex-row">
   {{> sidebar }}
   {{> projectsList}}
   {{> ../../partials/modals/projectForm}}
   {{> ../../partials/modals/warningModal}}
-</main>
